Add files via upload
This is R code that allows the user to create a function. The function mean() already exists but we created a function showing how the mean is calculated for the specific vector assignment2. The function we created is called MyMean and it calculates the sum of assignment2 and then divides it by the length of assignment2 returning the solution of the equation. The return function is needed to specify what value the function should return.