Add files via upload
I first created the three variables that were included in the assignment. I made the "Name" vector which shows the names of the candidates, and then the "ABCPoll" and "CBSPoll" which shows the political poll results from ABC and CBS for each of the seven candidates listed in the "Name" vector. I used the data.frame() function to put all the data into a dataset naming it "PollData". This function allowed me to organize all the data by placing it in rows and columns showing the number of polls each candidate received from ABC and CBS. Based on the dataset, Donald had the highest poll results out of all the candidates. He had 62 votes with ABC and 75 votes with CBS.