Add files via upload
In this assignment, I first created the variables of the dataset and put it in a dataframe using the data.frame() function. I was then able to create the boxplot and histogram regarding the patient's blood pressure and initial ratings using ggplot2. The blood pressure ranges from 32 to 205 showing high variability. According to the histogram most patients have moderate blood pressure values. However, there are a few extreme values that are shown within the range of 170 and 210 indicating potential outliers. The histogram otherwise shows that the blood pressure values are evenly distributed in the dataset. The boxplot shows the relationship between blood pressure and first assessments. According to the boxplot, patients who received a 1 (good) first assessment showed a higher median BP while patients rated a 0 (bad) showed a lower median in BP.  In conclusion, high BP is linked to "good" assessments and a lower BP is linked to "bad" assessments.