Add files via upload
I first created the two matrices that the assignment listed. The next step I took was to find the determinants of matrix A and matrix B. To do this I used the det() function. The determinant of matrix A was zero. I received an error when running det(B). This occurred because matrix B isn't a square matrix and contains 1000 elements which doesn't fit a 10 x 10 matrix. In conclusion, matrix A has no inverse because det(A) = 0, therefore the function solve(A) will not work. To produce the same result for matrix B, I would increase the number of elements to 10000 and make nrow = 100.